+++ Plugins/Phoenix/BookmarkService.swift
@@ -15,7 +15,7 @@ internal class BookmarkService {
                                    partnerId: Int,
                                    ks: String,
                                    eventType: String,
-                                   currentTime: Float,
+                                   currentTime: Int32,
                                    assetId: String,
                                    fileId: String) -> KalturaRequestBuilder? {
         
@@ -45,7 +45,7 @@ internal class BookmarkService {
 
     }
 
-    private static func createBookmark(eventType: String, position: Float, assetId: String, fileId: String) -> JSON {
+    private static func createBookmark(eventType: String, position: Int32, assetId: String, fileId: String) -> JSON {
         var json: JSON = JSON.init(["objectType": "KalturaBookmark"])
         json["type"] = JSON("media")
         //json["id"] = JSON(assetId)
